hangdire backgroun oluşturuldu senaryo bölümleri tamamlandı

+++ DTOs/ScenarioDto/CreateScenarioDto.cs
@@ -8,7 +8,7 @@ public sealed record CreateScenarioDto(
     string? ScenarioDescription,
     Guid? TriggerSensorId,
     TriggerTypeEnum TriggerType,
-    decimal? TriggerValue,
+    double? TriggerValue,
     DateTime? TriggerTime,
     Guid ActionSensorId,
     ActionTypeEnum ActionType,
